Move test_hermevander to separate files
